Offset & stroke size changes
The problem with the lines passing the nodes is related to the offset between lines (line 70). I changed the offset, but the lines still pass the nodes and the padding between nodes is a little too small. I also changed the line stroke on mouse over to try to accommodate this change.